+++ src/Service/RevisionSaver.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mediawiki\Api\Savers;
+namespace Mediawiki\Api\Service;
 
 use Mediawiki\Api\MediawikiApi;
 use Mediawiki\DataModel\Revision;