+++ src/Service/UserRepo.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Mediawiki\Api\Repos;
+namespace Mediawiki\Api\Service;
 
 use Mediawiki\Api\MediawikiApi;
 use Mediawiki\DataModel\User;
 
-class ApiUserRepo {
+class UserRepo {
 
 	/**
 	 * @var MediawikiApi